Path finding on a grid world
This code uses a graph class to encapsulate the implementation of the
graph behind functions like get_neighbors and cost so the we are not
tied to my decision to use a List[List[int]] to represent the grid.
There is then breath first search, dijkstras, and a star seach algorithms
implemented. One thing I don't love is my handling of the start and end
locations the idea is that the grid might not have them marked, it is
just a map you need to navigate through so I didn't want the grid reader
to return start and ends. I also didn't want the graph object to maintain
references to them because they aren't intrinsic properties of the graph.
I ended up creating a method on the graph that could find a node based on
its label. This isn't the greatest SoC because the graph shouldn't really
need to do this but the other option of making a function was hard because
the find is dependent on the implementation of the graph so being a method
lets us deal with that well. There might be a more elegant way to do this
but I haven't thought of a better way to do this yet.