Move Diagonally and Fuel Constrained Path Finding
Added the ability to move diagonally through the grid. We are still on a
grid world so the cost of moving up, down, left, or right is one. Moving
diagonally on a grid would be moving along the hypotenuse of a 45, 45,
90 triangle so the cost is √2

Added the ability to path-find with fuel constraints. This is done by
running multiple passes of the path finding algorithms. First we path
find from the start to all stations (and the end) and between all the
stations (and the end from there too). We then filter out the legs of
the paths that are greater than the agents capacity. The valid legs are
then used to create a meta graph (this time represented by nodes and
edges rather than a graph) and we re-run the path finding algorithms on
that graph. The selected steps in this meta-graph are then converted
back into their original grid path representation and combined into our
final path.